BeanEntity:\n 1.change int to Integer.\n 2.change long to Long

+++ core-parent/admin-core-biz/src/main/java/com/kalix/admin/core/biz/UserBeanServiceImpl.java
@@ -112,7 +112,7 @@ public class UserBeanServiceImpl extends ShiroGenericBizServiceImpl<IUserBeanDao
             organizationUserBean.setUpdateById(shiroService.getCurrentUserId());
             organizationUserBean.setUpdateBy(shiroService.getCurrentUserRealName());
             organizationUserBean.setOrgId(id);
-            organizationUserBean.setUserId(Integer.valueOf(jsonStatus.getTag()));
+            organizationUserBean.setUserId(Long.valueOf(jsonStatus.getTag()));
             organizationUserBeanDao.save(organizationUserBean);
         }
